+++ src/server/consts.ts
@@ -2,5 +2,6 @@ export const RDS_PREFIX = "@rds";
 export const RDS_CLIENT = `${RDS_PREFIX}/client`;
 export const RDS_OPEN_IN_EDITOR = `${RDS_PREFIX}/open-in-editor`;
 export const RDS_DEVTOOLS_UPDATE = `${RDS_PREFIX}/devtools-update`;
+export const FS_PREFIX = "@fs";
 export const DEPENDENCY_PREFIX = "@dep";
 export const ENTRY_POINT = "src/index.tsx";